requirements.txt
django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1
python-dotenv==1.0.0
asgiref==3.8.1
redis==7.1.0
dotenv==1.0.1